requirements.txt
alembic==1.9.2
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==1.2.0
dj-rest-auth==2.2.5
Django==3.2.18

django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0



django-rest-passwordreset==1.3.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.19
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.9.0
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
keyring==23.13.1
kombu==5.2.4
Mako==1.2.4
MarkupPy==1.14



oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3

Pillow==9.4.0
pkginfo==1.9.3
platformdirs==2.6.2
prompt-toolkit==3.0.33
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydot==1.4.2
Pygments==2.14.0
PyJWT==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.3
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.0.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
sqlparse==0.4.2
tablib==3.2.1
trio==0.22.0
trio-websocket==0.10.2
twine==4.0.2
types-cryptography==3.3.23
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0